+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.4.57](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.4.56...@codecademy/gamut-kit@0.4.57) (2021-06-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.4.56](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.4.55...@codecademy/gamut-kit@0.4.56) (2021-06-28)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.4.55](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.4.54...@codecademy/gamut-kit@0.4.55) (2021-06-28)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.4.54](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.4.53...@codecademy/gamut-kit@0.4.54) (2021-06-28)
 
 **Note:** Version bump only for package @codecademy/gamut-kit